Summary: Fixed several problems with maHttpGetResponseHeader on Android.
1. When giving maHttpGetResponseHeader a buffer with size 0, an exception occurred in the Android runtime instead of returning the size of the response header, according to the specification
2. Made the headers provided by maHttpGetResponseHeader null terminated according to specfication.
3. Fixed so that the JNI maHttpGetResponseHeader takes an integer address instead of a long address (like all other JNI syscalls). Since when it had a long address, the bufSize argument always became 0 on the Java side. I am not sure why, but it seems like the bufSize was coalesced with the buffer address.

+++ runtimes/java/platforms/androidJNI/AndroidProject/src/com/mosync/internal/android/MoSyncThread.java
@@ -2426,7 +2426,7 @@ public class MoSyncThread extends Thread
 	}
 	
 	int maHttpGetResponseHeader(
-		int connHandle, String key, long address, int bufSize)
+		int connHandle, String key, int address, int bufSize)
 	{
 		return mMoSyncNetwork.maHttpGetResponseHeader(
 			connHandle, key, address, bufSize);
